requirements.txt
py-redis
pymongo
dnspython
discord.py
dotenv
aio_pika